moving fee assertion logic to utils
